requirements.txt
streamlit
tensorflow>=2.8.0
numpy
pillow
scikit-learn
matplotlib
pandas